requirements.txt
torch>=1.11.0
torchvision>=0.12.0
pillow>=8.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
timm>=0.6.0
gdown>=4.0.0